requirements.txt
allure-pytest
pytest
faker
